refactor(task-status-sequencer): apply shared kernel pattern, move infra layer logic to root platform/obsidian, generalising for future modules.
