+++ src/platform/obsidian/editor/scroll-preserver.ts
